requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.10.4
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.10.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudmersive-validate-api-client==3.2.5
cloudmersive-virus-api-client==3.1.1
colorama==0.4.6
Deprecated==1.2.14
dnspython==2.3.0
email-validator==2.0.0.post2
et-xmlfile==1.1.0
Flask==2.3.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
flask-paginate==2022.1.8
Flask-Principal==0.4.0
Flask-ReCaptcha==0.4.2
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.3
flask-talisman==1.0.0
Flask-WTF==1.1.1
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.5.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mdurl==0.1.2
multidict==6.0.4
numpy==1.25.0
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.1
pandas==2.0.2
primePy==1.3
psycopg2-binary==2.9.6
Pygments==2.15.1
PyJWT==2.7.0
pyotp==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
rich==13.4.2
six==1.16.0
SQLAlchemy==2.0.13
stripe==5.4.0
twilio==8.2.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.4
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
